+++ examples/data-sources/iosxe_flow_record/data-source.tf
@@ -1,0 +1,3 @@
+data "iosxe_flow_record" "example" {
+  name = "FNF1"
+}