+++ examples/resources/iosxe_flow_record/import.sh
@@ -1,0 +1,1 @@
+terraform import iosxe_flow_record.example "Cisco-IOS-XE-native:native/Cisco-IOS-XE-flow:flow/record=FNF1"